deleted authentications and altered Product and ProductAPI

+++ Application/DTOs/ProductDTOs.cs
@@ -3,13 +3,13 @@ namespace Application.DTOs;
 
 public class PostProductDTO
 {
-    public int Price { get; set; }
+    public double Price { get; set; }
     public string Name { get; set; }
 }
 
 public class PartialUpdateProductDTO
 {
-    public int? Price { get; set; }
+    public double? Price { get; set; }
     public string? Name { get; set; }
     public int Id { get; set; }
 }
